feat: Add Redis utility for connecting to Redis server
This commit adds a new file `redis.ts` in the `src/utils` directory. The file contains code for connecting to a Redis server using the `ioredis` library. It also includes a listener for handling connection errors and logging them using the `logger` utility.

The Redis connection URL is fetched from the `constants.ts` file. The TLS options for the connection are currently commented out but can be enabled if needed.

This utility will be used for interacting with Redis in the application.